Move "capture warning" test helper to its own module
